pyproject.toml
[project]
name = "p-completeness-compendium-scripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-frontmatter>=1.1.0",
]

[tool.uv]
dev-dependencies = []
